Added files via upload
                                                                                 Search of Spatio-temporal Resources.
Background
This project evaluates algorithms on spatio-temporal data. The data in this case represents parking availability in a city.
Finding an on street parking spot in urban areas at peak hours is often not an easy task. Fortunately, detection technologies are advancing, which makes it possible to detect the availability of certain parking spots in some cases. SFpark  is such an example. SFpark is a project initiated by the San Francisco Municipal Transportation Agency (SFMTA), in which wireless sensors were embedded under the pavement of street parking spots to detect their occupancies on a real-time basis. The data of per-block availability (the number of available parking spots on each block) was published in real time. Knowing this information, it should make it easier to find an available parking spot as fast as possible. In general, finding a parking spot a